pyproject.toml
[tool.poetry]
name = "exchange-django"
version = "0.1.0"
description = ""
authors = ["aslanahmtv <aslanahmtv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django = "^5.1.5"
psycopg2 = "^2.9.10"
python-environ = "^0.4.54"
requests = "^2.32.3"
datetime = "^5.5"
openpyxl = "^3.1.5"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.8"
pyyaml = "^6.0.2"
uritemplate = "^4.1.1"
inflection = "^0.5.1"
django-rest-swagger = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
